requirements.txt
streamlit
langchain-core
langchain-community
langchain-openai
langchain-ollama
youtube-transcript-api
pytube
python-dotenv
openai
pillow
requests